+++ main.cpp
@@ -36,7 +36,7 @@
 #include "Abstract_ExpManager.h"
 #include "ExpManager.h"
 
-double staticMTime = 0;
+long long staticMTime = 0;
 void print_help(char* prog_path) {
     // Get the program file-name in prog_name (strip prog_path of the path)
     char* prog_name; // No new, it will point to somewhere inside prog_path
